perf(python): stop normalizing values that are already in 32-bit range
Int32 is now a plain Python `int`, so `int32()` went from being a type
constructor to being a Python-level function call. Call sites that were
free under the old representation are now two interpreter frames each.

On a property-based test suite (Scriptorium's Hedgehog, 21 tests), the
plain-int branch called `int32()` 1,346,191 times where 5.13.0 called it
zero times, and ran ~4% slower end to end despite the arithmetic itself
getting faster. Three of the hottest callers normalize nothing:

- `compare_primitives` wrapped the literals `0`, `-1` and `1`, which are
  in range by construction. 161k calls, the single hottest site.
- `@returns(int32)` on `compare` / `compare_arrays` / `compare_dicts`
  cost a `functools.wraps` wrapper plus an `int32()` on the library's
  hottest generic path. Every branch of these returns a literal or an
  already-normalized `compare`, except a user-supplied `CompareTo` --
  so normalize just that one, and drop the decorator. 86k calls.
- `int32()`'s own slow path recursed as `int32(int(value))`, re-testing
  `type(value) is int` on a value it already knows is one. 195k calls.

Also unwraps the remaining in-range integer literals across the library.

Together this drops `int32()` calls to 903,505 and takes the branch from
+4.1% to +1.9% against 5.13.0.

Adds coverage for the two places where semantics could have regressed:
`int32` still wrapping through the now-inlined slow path, and `compare`
still normalizing a `CompareTo` that returns an out-of-range value.

+++ src/fable-library-py/fable_library/diagnostics.py
@@ -49,7 +49,7 @@ class StopWatch:
 
 
 def frequency() -> int:
-    return int32(1000000000)
+    return 1000000000
 
 
 def get_timestamp() -> int64:
